+++ CHANGELOG.md
@@ -13,7 +13,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## `3.0.0` (2018-05-09)
 
 - Simplify the code to use an editor command
-- Contribute a context menu command *Draw a box*
+- Contribute a context menu command *Draw a box enclosing selection*
+- Contribute a new context menu command *Insert a drawing canvas* which inserts a 90 characters by 30 lines canvas of spaces
+- Bind `Ctrl/Cmd+D+B` to drawing a box and `Ctrl/Cmd+I+C` to inserting the canvas
 
 ## `2.0.0` (2018-04-11)